Completed HelpRequestPicture parts.

+++ Help.Domain.Services/HelpServiceAgg/HelpRequestPictureService.cs
@@ -2,7 +2,10 @@
 using Base_Framework.General;
 using Help.Domain.Core.HelpServiceAgg.Data;
 using Help.Domain.Core.HelpServiceAgg.DTOs.HelpRequestPicture;
+using Help.Domain.Core.HelpServiceAgg.DTOs.HelpServicePicture;
 using Help.Domain.Core.HelpServiceAgg.Services;
+using static Microsoft.EntityFrameworkCore.DbLoggerCategory.Database;
+using System.IO;
 
 namespace Help.Domain.Services.HelpServiceAgg
 {
@@ -18,34 +21,76 @@ namespace Help.Domain.Services.HelpServiceAgg
 
         public async Task<OperationResult> Create(CreateHelpRequestPictureDTO command, CancellationToken cancellationToken)
         {
+            var operation = new OperationResult(_type, 0);          
+
+            try
+            {
+                #region Save picture
+
+                command.Name = NameGenarator.GenerateUniqeCode() + Path.GetExtension(command.Picture.FileName);
+                string path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "Images", "HelpRequestPictures", command.Name);
+                FileHandler.SaveImage(path, command.Picture);
+
+                #endregion
+
+                int id = await _helpRequestPictureRepository.Create(command, cancellationToken);
+                await _helpRequestPictureRepository.Save(cancellationToken);
+                operation.RecordReferenceId = id;
+                return operation.Succedded();
+            }
+            catch
+            {
+                return operation.Failed(ApplicationMessages.CreationFailed);
+            }
+        }
+
+
+        public async Task<OperationResult> CreateDefault(CreateHelpRequestPictureDTO command, CancellationToken cancellationToken)
+        {
             var operation = new OperationResult(_type, 0);
+            command.Alt = "دیفالت ";
+            command.Title = "دیفالت ";
+            command.Name = "DefaultHelpServicePicture.jpg";
 
-            #region Save picture
-
-            command.Name = NameGenarator.GenerateUniqeCode() + Path.GetExtension(command.Picture.FileName);
-            string path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "Images", "HelpRequestPictures", command.Name);
-            FileHandler.SaveImage(path, command.Picture);
-
-            #endregion
-
-            await _helpRequestPictureRepository.Create(command, cancellationToken);
-
-            return operation.Succedded();
+            try
+            {
+                int id = await _helpRequestPictureRepository.Create(command, cancellationToken);
+                await _helpRequestPictureRepository.Save(cancellationToken);
+                operation.RecordReferenceId = id;
+                return operation.Succedded();
+            }
+            catch
+            {
+                return operation.Failed(ApplicationMessages.CreationFailed);
+            }
         }
 
         public async Task<OperationResult> Edit(EditHelpRequestPictureDTO command, CancellationToken cancellationToken)
         {
+            var operation = new OperationResult(_type, command.Id);
+
+            if (!await _helpRequestPictureRepository.IsExist(x => x.Id == command.Id, cancellationToken))
+                return operation.Failed(ApplicationMessages.RecordNotFound);
+
+
             if (command.Picture != null)
             {
                 string path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "Images", "HelpRequestPictures");
 
-                #region Delete Old Image
 
-                 FileHandler.DeleteFile(Path.Combine(path, command.Name));
+                if (command.Name != "DefaultHelpServicePicture.jpg")
+                {
 
-                #endregion
+                    #region Delete Old Image
 
-                #region Save picture
+
+                    FileHandler.DeleteFile(Path.Combine(path, command.Name));
+
+                    #endregion
+
+                }
+
+                #region Save
 
                 command.Name = NameGenarator.GenerateUniqeCode() + Path.GetExtension(command.Picture.FileName);
                 path = Path.Combine(path, command.Name);
@@ -54,11 +99,6 @@ namespace Help.Domain.Services.HelpServiceAgg
                 #endregion
 
             }
-
-            var operation = new OperationResult(_type, command.Id);
-
-            if (!await _helpRequestPictureRepository.IsExist(x => x.Id == command.Id, cancellationToken))
-                return operation.Failed(ApplicationMessages.RecordNotFound);
 
             await _helpRequestPictureRepository.Edit(command, cancellationToken);
             await _helpRequestPictureRepository.Save(cancellationToken);
